+++ injectool/__init__.py
@@ -1,9 +1,8 @@
 """Dependency injection tool"""
 
-__version__ = '2.0.4'
+__version__ = '3.0.0'
 
-from .core import DependencyError, Resolver, Container
-from .core import get_dependency_key, set_container, get_container, resolve, use_container
-from .resolvers import SingletonResolver, FunctionResolver
-from .resolvers import add_resolver, add_singleton, add_function_resolver, add_type
-from .injection import inject, dependency
+from .core import Dependency, Resolver, DependencyError, Container
+from .core import set_default_container, get_container, resolve, use_container
+from .resolvers import add, add_singleton, add_type, add_scoped, add_per_thread, scope
+from .injection import inject, dependency, In